fix: 나뉘어져 있던 ErrorCode 통합하여 구현 (#44)
