+++ coupon/src/main/java/com/sparta/hotdeal/coupon/application/dto/ErrorResponseDto.java
@@ -1,27 +1,21 @@
 package com.sparta.hotdeal.coupon.application.dto;
 
-import com.sparta.hotdeal.coupon.infrastructure.exception.ApplicationErrorCode;
-import java.time.LocalDateTime;
+import com.sparta.hotdeal.coupon.application.exception.ErrorCode;
+import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
+import java.time.LocalDateTime;
+
 @Getter
-@AllArgsConstructor
+@AllArgsConstructor(access = AccessLevel.PRIVATE)
 public class ErrorResponseDto {
 
     private LocalDateTime timestamp;
-    private Integer code;
+    private String code;
     private String message;
 
-    public ErrorResponseDto(ApplicationErrorCode errorcode) {
-        this(LocalDateTime.now(), errorcode.getCode(), errorcode.getMessage());
-    }
-
-    public ErrorResponseDto(String message) {
-        this(LocalDateTime.now(), ApplicationErrorCode.INTERNAL_SERVER_EXCEPTION.getCode(), message);
-    }
-
-    public ErrorResponseDto(ApplicationErrorCode errorcode, String message) {
-        this(LocalDateTime.now(), errorcode.getCode(), message);
+    public ErrorResponseDto(ErrorCode errorCode, String customMessage) {
+        this(LocalDateTime.now(), errorCode.getCode(), customMessage != null ? customMessage : errorCode.getDescription());
     }
 }
