--- coupon/src/main/java/com/sparta/hotdeal/coupon/infrastructure/exception/ApplicationException.java
@@ -1,11 +1,0 @@
-package com.sparta.hotdeal.coupon.infrastructure.exception;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public class ApplicationException extends RuntimeException {
-
-    public ApplicationErrorCode errorCode;
-}
